+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IDeliver.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IDeliver
     {
         [JsonProperty("deliver")]
-        int Deliver { get; set; }
+        long Deliver { get; set; }
 
         [JsonProperty("deliver_details")]
         RateDetails DeliverDetails { get; set; }
