+++ RabbitMqHttpApiClient/Models/QueueModel/QueueMessageStats.cs
@@ -5,10 +5,10 @@ namespace RabbitMqHttpApiClient.Models.QueueModel
 {
     public class QueueMessageStats : MessageStats, IDiskInfo
     {
-        public int DiskReads { get; set; }
+        public long DiskReads { get; set; }
         public RateDetails DiskReadsDetails { get; set; }
 
-        public int DiskWrites { get; set; }
+        public long DiskWrites { get; set; }
         public RateDetails DiskWritesDetails { get; set; }
     }
 }
